+++ common/src/main/java/dev/gimme/netherreset/domain/util/Constants.java
@@ -7,6 +7,6 @@ public class Constants {
 
     public static final String MOD_ID = "netherreset";
     public static final String MOD_NAME = "Nether Reset";
-    public static final String PLAYER_DIM_INVENTORY_ID = "player_dim_inventory";
+    public static final String PLAYER_DIM_INVENTORY_ID = "dim_inventory";
     public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
 }