+++ fabric/src/main/java/dev/gimme/netherreset/fabric/mixin/MixinPlayerDie.java
@@ -1,0 +1,22 @@
+package dev.gimme.netherreset.fabric.mixin;
+
+import dev.gimme.netherreset.Main;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.damagesource.DamageSource;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+/**
+ * Lets the mod know when a player dies.
+ */
+@Mixin(ServerPlayer.class)
+public class MixinPlayerDie {
+
+    @Inject(method = "die", at = @At(value= "HEAD"), require = 1)
+    private void onPlayerDeath(DamageSource damageSource, CallbackInfo ci) {
+        ServerPlayer instance = (ServerPlayer) (Object) this;
+        Main.INSTANCE.getPlayerHandler().onPlayerDeath(instance);
+    }
+}